+++ src/Controller/UserController.php
@@ -52,9 +52,9 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
-        if ($this->getUser()->getRoles()[0] != "ROLE_ADMIN")
-            if ($user->getIdCompany()->getId() != $this->getUser()->getIdCompany()->getId())
-                throw $this->createNotFoundException('The user does not exist');
+        if ($this->getUser()->getRoles()[0] != "ROLE_ADMIN" && $user->getIdCompany()->getId() != $this->getUser()->getIdCompany()->getId()) {
+            throw $this->createNotFoundException();
+        }
 
         return $this->render('user/show.html.twig', [
             'user' => $user,
@@ -64,6 +64,10 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser()->getRoles()[0] != "ROLE_ADMIN" && $user->getIdCompany()->getId() != $this->getUser()->getIdCompany()->getId()) {
+            throw $this->createNotFoundException();
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -82,6 +86,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser()->getRoles()[0] != "ROLE_ADMIN" && $user->getIdCompany()->getId() != $this->getUser()->getIdCompany()->getId()) {
+            throw $this->createNotFoundException();
+        }
+
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
